LibWeb: Move visual context tree construction to AVC module
ViewportPaintable owned both the current visual context tree and the
details for constructing it, keeping CSS transform, clip, and effect
derivation mixed into the viewport paint lifecycle.

Move the construction helpers into AccumulatedVisualContext.cpp and
leave ViewportPaintable responsible for owning the resulting snapshot.
The visual viewport context index is now local builder state instead of
viewport object state.
